+++ src/image_classifier/pytorch/__init__.py
@@ -1,0 +1,7 @@
+from .data_module import FashionMNISTDataModule
+from .lightning_module import FashionMNISTClassifier
+
+__all__ = [
+    "FashionMNISTDataModule",
+    "FashionMNISTClassifier",
+]